#5 Trick Crud with custom validators

+++ src/Controller/TrickController.php
@@ -13,14 +13,12 @@ use App\Repository\CommentRepository;
 use App\Repository\ImageRepository;
 use App\Repository\TrickRepository;
 use DateTimeImmutable;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\String\Slugger\SluggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Uid\Uuid;
-use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
@@ -33,9 +31,23 @@ class TrickController extends AbstractController
             'pageNb' => '\d+',
         ]
     )]
-    public function index(Request $request, TrickRepository $trickRepository, CommentRepository $commentRepository, EntityManagerInterface $entityManager, string $slug, Security $security, TranslatorInterface $translator, int $pageNb = 0): Response
+    /**
+     * index function display a detailled trick with comments. I ensure
+     *
+     * @param Request $request Permits to retrieve form data for comment creation.
+     * @param TrickRepository $trickRepository Necessary to fetch given trick by slug.
+     * @param CommentRepository $commentRepository necessary to fetch comment in bdd depending on how many comments there are and how many times loading more has click on.
+     * @param EntityManagerInterface $entityManager Tools to push to bdd.
+     * @param string $slug parameter to fetch the given trick.
+     * @param Security $security tools to set Author of the comment.
+     * @param integer $pageNb given page to compute how many comment to retrieve.
+     * @return Response
+     */
+
+
+    public function index(Request $request, TrickRepository $trickRepository, CommentRepository $commentRepository, EntityManagerInterface $entityManager, string $slug, Security $security, int $pageNb = 0): Response
     {
-        $trick = $trickRepository->getTrickBySlug($slug);
+        $trick = $trickRepository->findOneBy(['slug' =>$slug]);
         $commentList = $trick->getComments();
         $commentTotal = $trick->getComments()->count([]);
         $limit = 2;
@@ -64,25 +76,39 @@ class TrickController extends AbstractController
             $entityManager->persist($comment);
             $entityManager->flush();
 
-            $this->addFlash('success', $translator->trans('Trick.Comment.Done'));
+            $this->addFlash('success', ['Trick.Comment.Done']);
 
             return $this->redirectToRoute('app_trick_detail', ['slug' => $trick->getSlug()]);
         }
 
         return $this->render('trick/index.html.twig', [
-            'controller_name' => 'TrickController',
-            'trick' => $trick,
-            'commentList' => $commentList,
-            'form' => $form->createView(),
-            'pageNb' => $pageNb,
-            'commentTotal' => $commentTotal,
-            'limit' => $limit,
-            'limitReached' => $limitReached
+            'controller_name'       => 'TrickController',
+            'trick'                 => $trick,
+            'commentList'           => $commentList,
+            'form'                  => $form->createView(),
+            'pageNb'                => $pageNb,
+            'commentTotal'          => $commentTotal,
+            'limit'                 => $limit,
+            'limitReached'          => $limitReached,
         ]);
     }
 
     #[Route('/trickCreate', name: 'app_trick_create', methods: ['GET', 'POST'])]
     #[IsGranted('ROLE_USER')]
+    /**
+     * Creates a new trick.
+     *
+     * This method handles the creation of a new trick by a user. It processes the submitted form data, validates and saves the trick, and associates images and videos with it if provided.
+     *
+     * @param Request $request The HTTP request object containing form data.
+     * @param UploadImage $uploadImage An instance of the UploadImage service for handling image uploads.
+     * @param EntityManagerInterface $entityManager The Doctrine EntityManager to manage database operations.
+     * @param SluggerInterface $slugger The Slugger service to generate a unique slug for the trick.
+     *
+     * @return Response A Symfony Response object representing the rendered view or a redirection to the trick's detail page.
+     */
+
+
     public function trickCreate(Request $request, UploadImage $uploadImage, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
     {
         $trick = new Trick();
@@ -98,7 +124,6 @@ class TrickController extends AbstractController
             $trick->setUser($this->getUser());
             $trick->setSlug($slugger->slug($trickName));
             $images = $form->get('images')->getData();
-
 
             foreach ($images as $image) {
                 if ($uploadImage->validateUploadedFile($image)) {
@@ -118,12 +143,11 @@ class TrickController extends AbstractController
                 $video->setUuid(Uuid::v6());
             }
 
-            // VErification embeded link youtube Regex
-            // Error Message
+            // VErification embeded link youtube Regex.
             if (!$uploadImage->hasError()) {
                 $entityManager->persist($trick);
                 $entityManager->flush();
-                $this->addFlash('success', 'Trick.Created');
+                $this->addFlash('success', ['Trick.Created']);
                 return $this->redirectToRoute('app_trick_detail', ['slug' => $trick->getSlug()]);
             }
         }
@@ -141,14 +165,31 @@ class TrickController extends AbstractController
         methods: ['GET', 'POST'],
     )]
     #[IsGranted('ROLE_USER')]
+    /**
+     * Update an existing trick.
+     *
+     * This method handles the update of an existing trick by a user. It processes the submitted form data, validates and updates the trick, and associates images and videos with it if provided.
+     *
+     * @param Request $request The HTTP request object containing form data.
+     * @param TrickRepository $trickRepository The repository to fetch the existing trick by slug.
+     * @param EntityManagerInterface $entityManager The Doctrine EntityManager to manage database operations.
+     * @param string $slug The unique identifier (slug) of the trick to be updated.
+     * @param UploadImage $uploadImage An instance of the UploadImage service for handling image uploads.
+     * @param SluggerInterface $slugger The Slugger service to generate a unique slug for the trick.
+     *
+     * @return Response A Symfony Response object representing the rendered view or a redirection to the trick's detail page.
+     *
+     * @throws NotFoundHttpException if the specified trick does not exist.
+     */
+
+
     public function update(Request $request, TrickRepository $trickRepository, EntityManagerInterface $entityManager, string $slug, UploadImage $uploadImage, SluggerInterface $slugger): Response
     {
 
-        $trick = $trickRepository->getTrickBySlug($slug);
+        $trick = $trickRepository->findOneBy(['slug' =>$slug]);
         if (!$trick) {
             throw $this->createNotFoundException("This trick doesn't exist");
         }
-        $disable = false;
 
         $form = $this->createForm(TrickCreateFormType::class, $trick);
         $form->handleRequest($request);
@@ -183,7 +224,7 @@ class TrickController extends AbstractController
             if (!$uploadImage->hasError()) {
                 $entityManager->persist($trick);
                 $entityManager->flush();
-                $this->addFlash('success', 'Trick.Updated');
+                $this->addFlash('success', ['Trick.Updated']);
                 return $this->redirectToRoute('app_trick_detail', ['slug' => $trick->getSlug()]);
             }
         }
@@ -201,6 +242,23 @@ class TrickController extends AbstractController
         methods: ['GET', 'POST'],
     )]
     #[IsGranted('ROLE_USER')]
+    /**
+     * Set an image as the main image for a trick.
+     *
+     * This method allows a user to mark an image as the main image for a trick. The image with the specified UUID
+     * will be set as the main image, and the previously designated main image will be unmarked. After updating
+     * the main image, the user will be redirected to the trick update page.
+     *
+     * @param string $uuid The UUID of the image to set as the main image.
+     * @param ImageRepository $imageRepository The repository to fetch images.
+     * @param EntityManagerInterface $manager The Doctrine EntityManager for managing the database.
+     *
+     * @return Response A Symfony Response object representing a redirection to the trick update page.
+     *
+     * @throws NotFoundHttpException if the specified image does not exist.
+     */
+
+
     public function setMainImage($uuid, ImageRepository $imageRepository, EntityManagerInterface $manager)
     {
 
@@ -213,35 +271,105 @@ class TrickController extends AbstractController
         $manager->persist($image);
         $manager->flush();
         return $this->redirectToRoute('app_trick_update', ['slug' => $image->getTrick()->getSlug()]);
-
     }
 
     #[Route(
         '/delete/image/{uuid}',
         name: 'app_delete_image',
-        methods: ['DELETE'],
+        methods: ['POST', 'DELETE'],
     )]
     #[IsGranted('ROLE_USER')]
-    public function deleteImage($uuid, ImageRepository $imageRepository, EntityManagerInterface $manager)
+    /**
+     * Delete an image from a trick.
+     *
+     * This method allows a user to delete an image associated with a trick using its UUID. It first verifies
+     * the CSRF token provided in the request to prevent unauthorized deletions. If the token is valid,
+     * the image file is deleted, the image entity is removed from the database, and the user is redirected
+     * to the trick update page.
+     *
+     * @param Request $request The request object used for CSRF token validation.
+     * @param string $uuid The UUID of the image to delete.
+     * @param ImageRepository $imageRepository The repository to fetch images by UUID.
+     * @param EntityManagerInterface $manager The Doctrine EntityManager for managing the database.
+     *
+     * @return Response A Symfony Response object representing a redirection to the trick update page after image deletion.
+     *
+     * @throws NotFoundHttpException if the specified image does not exist.
+     * @throws AccessDeniedException if the CSRF token is invalid.
+     */
+
+
+    public function deleteImage(Request $request, $uuid, ImageRepository $imageRepository, EntityManagerInterface $manager)
     {
+
         $image = $imageRepository->findOneBy(['uuid' => $uuid], []);
 
-        unlink('uploads/image/' . $image->getFileName());
+        $data = $request->get('_token');
 
-        $manager->remove($image);
-        $manager->flush();
-        return $this->redirectToRoute('app_trick_update', ['slug' => $image->getTrick()->getSlug()]);
+        if ($this->isCsrfTokenValid('delete' . $image->getUuid(), $data)) {
+            unlink('uploads/image/' . $image->getFileName());
+
+            $manager->remove($image);
+            $manager->flush();
+
+            $this->addFlash('success', 'Image.Deleted');
+
+            return $this->redirectToRoute('app_trick_update', ['slug' => $image->getTrick()->getSlug()]);
+        }
+
+        $this->addFlash('error', ['Token.Error']);
+        return $this->redirectToRoute('app_homepage');
     }
 
     #[Route(
         '/trick/delete/{slug}',
-        name: 'app_trick_delete',
-        methods: ['GET', 'POST'],
+        name: 'app_delete_trick',
+        methods: ['POST', 'DELETE'],
     )]
     #[IsGranted('ROLE_USER')]
-    public function deleteTrick($slug, EntityManagerInterface $manager, TrickRepository $trickRepository)
+    /**
+     * Delete a trick.
+     *
+     * This method allows a user to delete a trick based on its slug. It verifies the action is allowed by validating
+     * the CSRF token provided in the request. If the token is valid, it fetches the files linked to the trick and unlinks them.
+     * After unlinking the files and removing the trick entity from the database, the user is redirected to the homepage.
+     *
+     * @param Request $request The request object used for CSRF token validation.
+     * @param string $slug The slug of the trick to delete.
+     * @param EntityManagerInterface $manager The Doctrine EntityManager for managing the database.
+     * @param TrickRepository $trickRepository The repository to fetch tricks by slug.
+     *
+     * @return Response A Symfony Response object representing a redirection to the homepage after trick deletion.
+     *
+     * @throws NotFoundHttpException if the specified trick does not exist.
+     * @throws AccessDeniedException if the CSRF token is invalid.
+     */
+
+
+    public function deleteTrick(Request $request, $slug, EntityManagerInterface $manager, TrickRepository $trickRepository)
     {
+        $data = $request->get('_token');
+        $trick = $trickRepository->findOneBy(['slug' =>$slug]);
+        if (!$trick) {
+            throw $this->createNotFoundException("This trick doesn't exist");
+        }
 
+
+        if ($this->isCsrfTokenValid('delete' . $trick->getSlug(), $data)) {
+            foreach ($trick->getImages() as $image) {
+                unlink('uploads/image/' . $image->getFileName());
+            }
+
+            $manager->remove($trick);
+            $manager->flush();
+
+            $this->addFlash('success', ['Trick.Deleted']);
+
+            return $this->redirectToRoute('homepage');
+        }
+
+        $this->addFlash('error', ['Token.error']);
+
+        return $this->redirectToRoute('homepage');
     }
-
 }
